cheri: Fix compile in drivers/crypto/
- Standard uintpr_t vs. unsigned long issues.
- Adapt to changed tasklet calling conventions.
- Increase alignment in drivers/crypto/intel/qat/qat_common/adf_cfg_user.h
- Pointer handling in drivers/crypto/intel/qat/ is somewhat dubious.

+++ drivers/crypto/amcc/crypto4xx_core.c
@@ -1070,7 +1070,7 @@ static void crypto4xx_unregister_alg(struct crypto4xx_device *sec_dev)
 	}
 }
 
-static void crypto4xx_bh_tasklet_cb(unsigned long data)
+static void crypto4xx_bh_tasklet_cb(uintptr_t data)
 {
 	struct device *dev = (struct device *)data;
 	struct crypto4xx_core_device *core_dev = dev_get_drvdata(dev);
@@ -1419,7 +1419,7 @@ static int crypto4xx_probe(struct platform_device *ofdev)
 
 	/* Init tasklet for bottom half processing */
 	tasklet_init(&core_dev->tasklet, crypto4xx_bh_tasklet_cb,
-		     (unsigned long) dev);
+		     (uintptr_t) dev);
 
 	core_dev->dev->ce_base = of_iomap(ofdev->dev.of_node, 0);
 	if (!core_dev->dev->ce_base) {
