+++ drivers/crypto/atmel-aes.c
@@ -2165,14 +2165,14 @@ static void atmel_aes_dma_cleanup(struct atmel_aes_dev *dd)
 	dma_release_channel(dd->src.chan);
 }
 
-static void atmel_aes_queue_task(unsigned long data)
+static void atmel_aes_queue_task(uintptr_t data)
 {
 	struct atmel_aes_dev *dd = (struct atmel_aes_dev *)data;
 
 	atmel_aes_handle_queue(dd, NULL);
 }
 
-static void atmel_aes_done_task(unsigned long data)
+static void atmel_aes_done_task(uintptr_t data)
 {
 	struct atmel_aes_dev *dd = (struct atmel_aes_dev *)data;
 
@@ -2348,9 +2348,9 @@ static int atmel_aes_probe(struct platform_device *pdev)
 	spin_lock_init(&aes_dd->lock);
 
 	tasklet_init(&aes_dd->done_task, atmel_aes_done_task,
-					(unsigned long)aes_dd);
+					(uintptr_t) aes_dd);
 	tasklet_init(&aes_dd->queue_task, atmel_aes_queue_task,
-					(unsigned long)aes_dd);
+					(uintptr_t) aes_dd);
 
 	crypto_init_queue(&aes_dd->queue, ATMEL_AES_QUEUE_LENGTH);
