+++ drivers/crypto/omap-sham.c
@@ -1704,7 +1704,7 @@ static struct ahash_engine_alg algs_sha384_sha512[] = {
 },
 };
 
-static void omap_sham_done_task(unsigned long data)
+static void omap_sham_done_task(uintptr_t data)
 {
 	struct omap_sham_dev *dd = (struct omap_sham_dev *)data;
 	int err = 0;
@@ -2063,7 +2063,7 @@ static int omap_sham_probe(struct platform_device *pdev)
 	platform_set_drvdata(pdev, dd);
 
 	INIT_LIST_HEAD(&dd->list);
-	tasklet_init(&dd->done_task, omap_sham_done_task, (unsigned long)dd);
+	tasklet_init(&dd->done_task, omap_sham_done_task, (uintptr_t) dd);
 	crypto_init_queue(&dd->queue, OMAP_SHAM_QUEUE_LENGTH);
 
 	err = (dev->of_node) ? omap_sham_get_res_of(dd, dev, &res) :
